Update workers-script.js
Updated to add a leaderboard. The leaderboard uses a cloudflare key-value store for high frequency access to a JSON file, which is updated every time a team checks in (this might not be the most efficient thing to do for big events with high levels of check-ins). The leaderboard page uses the same event code as the team is using for the game, and updates every ~10 seconds.